requirements.txt
aiofiles==25.1.0
aiogram==3.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
async==0.6.2
asyncio==4.0.0
attrs==25.4.0
certifi==2026.2.25
colorama==0.4.6
decorator==5.2.1
deprecation==2.1.0
ExifRead==3.5.1
eyeD3==0.9.9
filetype==1.2.0
frozenlist==1.8.0
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
Kurigram==2.2.19
lxml==6.0.2
magic-filter==1.0.12
moviepy==2.2.1
multidict==6.7.1
numpy==2.4.3
packaging==26.0
pillow==11.3.0
proglog==0.1.12
propcache==0.4.1
pyaes==1.6.1
pydantic==2.12.5
pydantic_core==2.41.5
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.2.2
python-magic==0.4.27
telegram==0.0.1
TgCrypto==1.2.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.23.0
